Initial commit
I created the README.md document to document and describe the project
to developers that are new to the project. Over time, this document
will evolve into a getting started document for new developers
containing instructions on how to build the project from the source
code and how to contribute to the project.

I created the .gitignore and .gitattributes files from GitHub
templates.