Receive message from MSMQ queue and start a transaction
I created the DependentTransactionMessageQueueSourceBlock class. This
class will begin a transaction, receive a message from an MSMQ queue,
and will propagate the transaction through the pipeline to be completed
and committed later. I implemented the unit test code and supporting
classes for creating transactions and dependent transactions, and
receiving messages in a transaction.