Send message to MSMQ queue and end transaction
I created the DependentTransactionMessageQueueTargetBlock class. The
new class will send messages to an MSMQ queue as part of a transaction
and will complete the transaction. I implemented the unit tests for the
new class and enhanced the support services with new APIs needed to
send messages as part of transactions and to commit or rollback the
transactions.